add tests for quad9 fluid2 and inf-sup fluid2 (Taylor-Hood)
The odd thing is that the inf-sup version produces an floating point exception
(inside UMFPACK) if those are enabled (TRAP_FE defined) but works well without it. 
The exception vanishes if no pressure node is fixed. Umfpack works well in this
case but of course the pressure value does not converge. The velocity does.

I think the pressure singularities are to blame. These seem to demand a certain
pressure level and to overrule them hurts the direct solver.

aztec currently does not work with inf-sup elements.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@2553 060e22d8-9d1c-0410-868a-e3a950d63d95