Processor local binary output (for those who need it.) Simply set the PROCESSOR_LOCAL flag in the IO section of your input file to Yes. This will create extra output files for each processor. (So you end up with a lot of files.)
This way you can write to the node local scratch area. But even the
output to our NFS shared home directories should be much more robust.

As before there is one control file for the whole set. And as long as
all result files are in the same directory as the control file (you
will have to copy your files if you write to the scratch space on the 
nodes) the filters manage to read the results easily. Isn't it great?

Restart from processor local files is still missing.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@2715 060e22d8-9d1c-0410-868a-e3a950d63d95