The code is now assembling everything != 0.0 (in the past we had a "weak 0" corresponding to 10E-10)
The change is absolutely necessary for the stabilised fluid 
problem since some stabilisation terms --- including the 
(grad p, grad q) --- are scaled with dt^2. For smaller timesteps 
they may be in the magnitude of 10E-14 and have been omitted by the 
standard assembly procedure.
The consequence was an immediate numerical blow up of the solution
(changing time step by factor 10 changed the solution by around 10E+06...)






git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@2862 060e22d8-9d1c-0410-868a-e3a950d63d95