The Aztec_MSR solver option AZ_conv may now be read from the input file. The keyword is AZCONV.
The default value is AZ_none as it has been before.
If Aztec_MSR is applied to incremental solution algorithms --- for example
fluid problems with FLUID_INCREMENTAL define flag --- it might be useful to
switch to AZ_rhs or AZ_r0.

For a detailed overview over available options see the Aztec manual...






git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@2874 060e22d8-9d1c-0410-868a-e3a950d63d95