FSI specific matrix free Newton Krylov interface solver. semi-working version
We want to apply a matrix free Newton Krylov solver to the 
FSI interface residuum

	F(d) = -R(d) = d(i+1) - d(i) = S(F(d(i))) - d(i) = 0

That is we want to solve the equation

	R' x = -R

where we derivate R with respect to the interface displacements,
using an iterative solver that only requires the evaluation 
of R' x and does not need the Jacobian R' itself. This is
important as we cannot calculate R' but we can approximate
R' x. In general we have:

	R'   = S'(F(d(i))) F'(d(i))   - I

	R' x = S'(F(d(i))) F'(d(i)) x - x

To evaluate the second equation with given x we first have to
calculate

	z = F'(d(i)) x

which means to apply x to the derivative of the fluid field
residuum with respect to any interface displacements. That is
the derivation does include the effect of the fluid mesh 
displacement as well as the prescribed fluid interface velocity
(that in turn depends on the interface displacement.) We cannot
calculate this derivative. We can, however, neglect the mesh
displacement. In this case the above equation requires a
linear fluid solve with the prescribed interface displacement x.

In the same way the term

	y = S'(F(d(i))) z

requires a linear solve of the structure field. Finally

	R' x = y - x

is trivial.

The point is that both the (reinterpreted) fluid solver as well
as the structural solver are already implemented as part of
our steepest descent relaxation. So no substantial coding is
required to obtain a finite difference free (read parameter free)
solver for the FSI interface equation that belongs to the same
class of quasi-Newton methods as the one by Gerbeau and Vidrascu:
"A Quasi-Newton Algorithm Based on a Reduced Model for Fluid-Structure
Interaction Problems in Blood Flows".

The down point is that often enought the interface displacement x
is too big to maintain a reasonable fluid mesh. So the algorithm
turns out to be rather sensitive and not yet ready for use.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@2932 060e22d8-9d1c-0410-868a-e3a950d63d95