FSI specific matrix free Newton Krylov interface solver. working version
Since the operation

	y = S'(F(d)) F'(d) x

is linear, we can easily scale x so avoid fluid mesh failtures.
Right now we scale by 1/(1e4*|x|). This solves the issue for
at least one input file. :)

So there is a working FSI specific method. It is indeed closely
related to the known matrix free method based on a finite
difference. For small delta

                              R(d + x delta) - R(d)
                       R' x = ---------------------
                                      delta

can be reformulated

                           R(d) + linR(x delta) - R(d)
                    R' x = ---------------------------
                                      delta

where linR(x delta) represents the linear addition to R(d) and
thus has to be evaluated at d. Of course this comes down to

			R' x = linR(x)

and explains why there are no boundary conditions on any of the
fields behind linR(x) other than the interface displacement x.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@2933 060e22d8-9d1c-0410-868a-e3a950d63d95