Add an (unsuccessful) attempt to improve convergence speed of the nonlinear interface iteration via the extrapolation used by Deparis and Michler (in different settings).
The extrapolation might improve convergence for poor schemes. But even if it never
does I feel the code is still interesting enough to be kept.

Another open question is how to handle the convective velocity when calculating the
steepest descent relaxation parameter. The current implementation just uses the
mesh movement due to the interface perturbation (starting with an undeformed mesh.)
I think we should take the ordinary mesh deformation of the time step into account.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@2943 060e22d8-9d1c-0410-868a-e3a950d63d95