proposed reference for node numbering scheme within BACI
maybe it could be moved into the Element class or another place (Ulrich, Michalel?).

Consider this as a proposal for discussions (maybe in the next BACI meeting).
We need it for XFEM (lots of geometric functions), so all elements that shall be enriched in the future should have a consistent node numbering scheme for reference. It does not have to be the same as this one, but this is already used in Gid, Cubit and Paraview, so it is compatible to all linear and quadratic elements in BACI already.

If we could find a good way of sharing this information, things like shape functions, which correspond to a specific node number scheme could be shared as well. This would reduce code size and the posibilities of coding errors due to copying.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@3433 060e22d8-9d1c-0410-868a-e3a950d63d95