Progress on structural multi-scale analyses.
A second problem instance with its own discretization is used for a microscale REV at the macroscopic Gauss points.
The macroscopic deformation gradient is passed to the microscale to define boundary conditions of the REV. 
For solving the microscale problem and passing arguments, a class MicroStruGenAlpha is used.
A homogenization routine for the calculation of macroscopic stress and constitutive matrix is already present but 
currently does nothing else than passing standard St Venant Kirchhoff parameters.
The "history variables" are stored for each Gauss point in MicroMaterial which is derived from Material.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@3505 060e22d8-9d1c-0410-868a-e3a950d63d95