Drop Fedora Core 5 support because we do not use it any more. Add support for blitz++ library. (http://www.oonumerics.org/blitz)
blitz++ provides serial dense vectors and matrices (or arrays with
arbitrary dimension) with nice syntax and highest possible performance.
The library is meant to supplement Epetra_SerialDenseSomething.

For user documentation on blitz++ see 
/lnm/lib/fc6/lam/share/doc/blitz-0.9/blitz.html
Doxygen generated documentation is available as well
/lnm/lib/fc6/lam/share/doc/blitz-0.9/doxygen/index.html

blitz++ is installed in /lnm for use on desktop maschines as well as
on the cluster.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@3617 060e22d8-9d1c-0410-868a-e3a950d63d95