Reimplementation of Fluid3 element. As clean as possible and as fast as before.
This is a new approach to the fluid element that aims at a clean implementation.
The mathematical core of the element has been moved to an internal class
that is separated from the main fluid element class. This allows us to allocate
the element working arrays just once. Additionally we avoid the mess that is
caused by different fluid implementations in the same element.

One point here is the heavy use of blitz to simplify the element level
matrix expressions. This results in very readable code without impact
on execution speed.

Right now the original code is still there. Further cleanup in fluid3_evaluate.cpp
will follow.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@3645 060e22d8-9d1c-0410-868a-e3a950d63d95