minimal polynomial vector extrapolation method due to Avram Sidi to accelerate FSI coupling. Working version.
It is a Krylov subspace method on difference vectors (a kind of secant method.)
So the huge point is that no Jacobian is needed. On the other hand it is
numerically a little sensitive. An idea like that has been used by Jan Vierendeels
and Simone Deparis before. Even the infamous Michler coupling builds on this
idea (but does it the wrong way.) The point here, however, is a much more sane
implementation (due to Avram Sidi) and the reference to a bunch of mathematical
papers.

The method is able to crunch hard problems. It is unclear however, if it can
have an advantage over Irons & Tuck's version of Aitken acceleration. It
might have in some cases if the size of the krylov subspace, the tolerance
and the fixed relaxation parameter are carefully selected.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@3668 060e22d8-9d1c-0410-868a-e3a950d63d95