GCR and GMRES both work as solvers at the FSI interface if the FD based MFNK is used. GCR does not work with MFNK based on approximated fluid derivatives, that is if the Jacobian operator approximation is incomplete. GMRES works fine it that case, as already known.
The point is that GCR allows to keep the Krylov space (or parts of it) between
solver calls, something plain GMRES cannot do. It seems that you can gain
something by keeping old krylov vectors, even if the number of field iterations 
increases rather than decreases. The old vectors work like a preconditioner.
But then again it can hurt to keep those vectors. The point is, I guess, to
keep the right ones and not too many of them. It is still unknown how to
choose the ones to keep. But we do not expect a tremendous gain anyway.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@3709 060e22d8-9d1c-0410-868a-e3a950d63d95