Now there is the possibility to keep just the n most important krylov vectors, where the definition of "most important" follows the idea of C. Vuik, who uses those vectors that have the largest effect on the residual.
This selection, however, does not improve anthing. The difficulty seems to be
that the interface equation changes during the newton iteration and the krylov
spaces from the first linear solve simply do not match the later ones. So one
can observe a degeneration of the newton scheme and an increase of the number
of newton steps.

The acceleration stems from the linear solver inside the newton scheme. 
Here the old Krylov vectors decrease the number of iteration required. But it 
is dubious whether the acceleration will pay off in the long run.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@3710 060e22d8-9d1c-0410-868a-e3a950d63d95