New defines file for new discretization.
We support both the old ccarat code and baci, the new implementation, but both versions
require very different define flags. This is it about time to split the definition files.
The file config/defines now contains just those flags that have a meaning for ccarat.
To compile baci please use a copy of config/defines.drt.

For baci we use fewer flags. In particular we do not hide our algorithms anymore. So
the D_FLUID and D_FSI flags are gone, as are a lot of experimental flags we used
to keep around.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@3755 060e22d8-9d1c-0410-868a-e3a950d63d95