the D_XFEM flag is gone, now all routines are compiled as default
However, to reduce external dependency, a QHULL flag is introduced, which allows to switch on and off
the support for the QHULL library. It is needed for the computation of intersected elements.
If an XFEM simulation is started without this flag activated,
the code will proceed until it actually needs the library and then crashes with a helpful message ;-)

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@3800 060e22d8-9d1c-0410-868a-e3a950d63d95