added a little library to create an Gmsh string from Elements and IntegrationCells
it can be printed to screen or to file. In the latter case, the mesh can be viewed using Gmsh.

Maybe, if the functions mature, there is some use for this in the post filters.

Note: it's for debuging only (A.Gerstenberger/U.Mayer), so it works only for Hex8/Hex27 (hardcoded)
it could be modified any time without warning, so use it at your own risk.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@3936 060e22d8-9d1c-0410-868a-e3a950d63d95