Progress on multi-scale approach for solids.
Two versions of the homogenization procedure are now available, 
namely a static one (neglecting the inertial forces on the microscale)
and a dynamic one. Currently you have to set the desired alternative in 
src/drt_mat/micromaterial.cpp but a possibility to make that choice in
the inputfile will come soon. 
In the dynamic case, no constitutive tensor is calculated, thus eventually
you should use it in combination with a matrix free method. Since I am 
currently working on that the dynamic homogenization still works with a 
dummy (St. Venant Kirchhoff) constitutive tensor.

After having fixed some bugs ;-) the constitutive tensor determined in the
static version was confirmed to coincide with a reference solution for 
different test cases.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4098 060e22d8-9d1c-0410-868a-e3a950d63d95