little change in the naming of the post_drt filter executables:
when compiling the post_drt_ensight or post_drt_gid filter 
based on PARALLEL configuration, a "_par" suffix is added.
For serial configuration, the familiar naming without extension 
is kept for convenience (for the moment).

motivation of this preparative work:
At the moment, a real bottleneck in our simulation chain is filtering the BACI output 
to the desired format (Ensight/GiD). 
A serial filter version runs out of RAM for problems containing some millions of elements/dofs
(system administrators do not really like swapping on cluster machines!)
This is why a parallel version of post_drt_filter will be necessary for everyone dealing
with huge discretizations. 

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4246 060e22d8-9d1c-0410-868a-e3a950d63d95