Derived an XFEM Writer from EnsightWriter
on the way, several methods of EnsightWriter where reviewed and slightly modified.
The purpose is to have a clear separation between methods that write to files, update internals of the EnsightWriter and support methods that do not change EnsightWriter or write to file
This hopefully makes deriving new classes easier.

Finally, I lifted those methods from private to protected that do not modify the base class.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4303 060e22d8-9d1c-0410-868a-e3a950d63d95