Silenced warnings about uninitialized gaussrule variables:
Problem is that we call dserror() after switching over element types, if we don't want to define a gaussrule for certain elements. So we are sure, the programm stops, but the compiler doesn't know that -> a lot of warnings about uninitialized gaussrules.

Solution -> Introduced for 1d, 2d and 3d gaussrule an additional enum entry (like intrule3D_undefined).
it can be used to initialize the gaussrule variable, before we start a switch stament to determine the gaussrule.

The additional safety is that if we call the routines to fill the arrays with the weights and gausspoint positions, baci stops, since no rule is defined for intrule3D-undefined ( ;-) ).

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4383 060e22d8-9d1c-0410-868a-e3a950d63d95