Refactoring of monolithic FSI algorithm class and introduction of overlapping MFSI coupling (not yet functional)
The point here is to put the Thyra details and things that are common to all monolithic FSI solvers
in a common base class. Derived classes just have to specify how the field solvers are combined,
how convergance is tested and what preconditioner is to be used. That's what makes the FSI algorithm
work.

What is needed here are much more elaborated preconditioners and the calculation of the structural
residual at t(n+1).


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4462 060e22d8-9d1c-0410-868a-e3a950d63d95