Communication speedup
Pay with memory, gain speed... Change the exporter to send all objects
at the same time. This way the memory requirement increases, but we speed
up a lot!

The method DRT::Exporter::Export is far from perfect. In particular
the gathering of the packed objects is bad as we need to reallocate
the sendblock a lot. But there is nothing much we can do about that
other that (a) ask the parobject for its size a priori (which is not
imeplemented) or (b) pack each parobject twice (the first time just
to get its size). If memory was not a concern we could pack all
parobjects in advance, put the small char vectors into a list and
setup sendblock just once afterwards.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4497 060e22d8-9d1c-0410-868a-e3a950d63d95