creation of overlapping MFSI system by explicit matrix splitting and implicit combination
So here are two ideas. For now we do not touch the assembling, thus we need to split
the system matrices of the fields once we get them. We'll see whether this will be too
expensive.

The combination of these three 2x2 matrices requires the shift of the interface
matrices into the structural interface map. This is difficult to do, so we use
Thyra here and create an implicit CouplingOperator, that shifts the vectors
before and/or after they are applied internally. Shifting vectors can be done the
usual way using Coupling objects. (Maybe it will even be possible to use
mortar coupling this way.) 

The down side to this is, that we are constrained to use iterative solvers
on the shifted interface matrix. And maybe we will have difficulties to come up
with a reasonable preconditioner. (The idea is to use the structural matrix
here as it is unshifted.) The rectangular matrices are not used inside a
solver, so there should be no issue with them.

I hope these ideas work.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4508 060e22d8-9d1c-0410-868a-e3a950d63d95