Introduce ConditionDofMap class to FSI that represents a splitted map and continue overlapping MFSI
The point here is that we need a place to store the Epetra_Importers that
translate between a full map and part of a map (either interface part or
field internal part, we need both.) This is now done by the ConditionDofMap
class. The Coupling class uses the same maps as ConditionDofMap, thus there
is finally just one map for each interface. (The setup process still needs
work...)

This class hides the Epetra_Importers that tend to puzzle people. And
the overlapping MFSI can easily (and elegantly) be done this way.
Looks great.

Now the costomized overlapping MFSI preconditioner needs to be build.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4518 060e22d8-9d1c-0410-868a-e3a950d63d95