STRUCTURAL DYNAMICS W/ GENERALISED-ALPHA
changed convergence criteria back to
while (((norm_>toldisp && fresmnorm>toldisp) || volnorm > toldisp ) && numiter<=maxiter)

This means: The NRI is considered converged if _either_ the norm of the force residual is smaller than tolerance _or_ the norm of the iterative incremental displacements (or both).

An absolute check of the force residual does not make sense due to the floating point character of the number model with limited precision, ie if the magnitude of the force residual is far larger than the provided tolerance, the machine is unable to achieve the requested criteria. Thus, we need a relative convergence check rather than an absolute check which "almost never works". Lena is going to do it... otherwise spring 2008...

Subversive CommiTheme (TM):
"Less is more." (L. Mies-van-der-Rohe)



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4523 060e22d8-9d1c-0410-868a-e3a950d63d95