Add very simple (diagonal) preconditioner for overlapping MFSI
Here we have a bunch of fixes, an explicit function to extract epetra field vectors
from a thyra product vector (used with the vector of unknowns in Evaluate())
and the simple diagonal preconditioner.

The preconditioner needs more entries so that a triangular scheme is obtained.
And the convergance test needs work.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4528 060e22d8-9d1c-0410-868a-e3a950d63d95