refactor Export method and add ExporterHelper classes
This allows to use one generic export algorithm to work for
any types of objects. We reduce the amount of templated code
in the header file and get a system that is simple to extend.

The reason for this is the addition of a third export method
that works for any plain old data (POD) type.

  template<typename T> void Export(map<int,T>& data);

As usual in this file I left the old code in place.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4570 060e22d8-9d1c-0410-868a-e3a950d63d95