switch overlapping MFSI from implicit Thyra addition to explicit Epetra addition of interface equations
This is one step backward to familiar ground. Instead of splitting the structural matrix
(along with the fluid and ale matrices) I perform the addition at the interface explicitely
on Epetra matrices. This mandates a few heavy weight Epetra matrix communications. But this
way I can build the (first) preconditioner using direct solvers. Overall a huge simplification.

The main point here is that this way a Dirichlet-Neumann preconditioner can be build
easily. And the added fluid contributions on the structural matrix might help a lot.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4580 060e22d8-9d1c-0410-868a-e3a950d63d95