Introduce triangular preconditioning in overlapping monolithic FSI
Wow! It works! The number of linear iteration halves!

The original (simple) preconditioner was block diagonal. So it involved
the equations

	sy = S^-1 sx
	fy = F^-1 fx                                           (1)
	ay = A^-1 ax

where S, F, A are the system matrices of structure, fluid and ale.
The structural matrix including the interface dofs, the other two
without interface.

Now we introduce a triangular coupling like this.

	sy = S^-1 sx
	fy = F^-1 fx - F^-1 F(I,Gamma) S^-1 sx                 (2)
	ay = A^-1 ax - A^-1 A(I,Gamma) S^-1 sx

where F(I,Gamma) and A(I,Gamma) are the (rectangular) matrices that
couple the internal fluid and ale dofs to the structure interface.
Of course, we actually just solve the systems

	S sy = sx
	F fy = fx - F(I,Gamma) sy                              (3)
	A ay = ax - A(I,Gamma) ay

so the numerical work as compared to (1) hardly increases. And as
we move from block Jacobi (1) to block Gauss-Seidel (3), we get
twice the convergance rate.

Thyra rocks!


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4670 060e22d8-9d1c-0410-868a-e3a950d63d95