Move and rename FSI::ConditionDofMap to DRT::UTILS::MapExtractor
The MapExtractor is a utility class that splits a dof row map from a
DRT::Discretization into two submaps. Communication of Epetra_Vectors
bewteen those maps is supported.

This class is meant to be used each time you need to create subvectors.

A particular point here is the way the splitting condition is introduced
to MapExtractor. As the condition which dof to put into which map is
user specific, the user has to supply this knowledge. So we apply the
strategy pattern [Gamma:1995] and have the user supply a condition class
that does the decision. Examples of such classes are included.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4756 060e22d8-9d1c-0410-868a-e3a950d63d95