Better condition input with validation and dat file generation support to help exodus filter
The short message is you can get the list of all valid conditions via

	./cca_fc6_lam.debg -d

(along with the list of valid input parameters.) and there is just
one internal definition for each condition that is used for reading
and writing.

The long message contains a few more details.

It works just like it does for the input parameters. We have a function called 
ValidConditions() (defined in drt_validconditions.cpp) that builds the data structures 
that define what each condition looks like. This is the place to go if you want
to modify or add an input condition in baci. (The other place is lnm_general.gid, of
course.)

For each condition dat file section an object of type ConditionDefinition is
created. This object knowns the section name, the condition name and type,
the geometry and so on. It is the very object that reads and writes a condition.
To do so it needs to know the components of the condition input line.
Remember, in a dat file each condition is defined by one line. The line
always starts with

	E num -

where num is the number of the design object (nodal cloud) the condition 
applies to. These three tokens are followed by whatever the condition
requires.

The idea is that the rest of the condition input line can be split in 
a set of components, each one describing one variable to be set to the
DRT::Condition object. These variables are:

- single strings
- single integers
- single doubles
- integer vectors
- double vectors

In order to read these variable types, we add appropriate components using
the AddComponent() method of ConditionDefinition. The available component
classes are:

- StringConditionComponent
- IntConditionComponent
- RealConditionComponent
- IntVectorConditionComponent
- RealVectorConditionComponent

You can probably see the point. All that needs to be done is to build up
a description of the condition input line using component classes.
Afterwards the condition input comes for free. The DRT::Condition objects
will be created and set to your Discretization.

When compared to the situation up to now, we gain a huge unification and
the separation of reader details and condition definition. This is a good
thing. There are, however, a few things that are not possible anymore.
In particular non-contextfree actions are not possible. For this reason
the periodic boundary condition definition needs to include the "Master"
or "Slave" flag explicitly. Some sanity tests had to be removed as well
in that condition. (Note to self: Talk to Peter.)

This is a major cleanup. Expect and report bugs.

And at the very end the best news. :) There is a new, unvalidated, dat
file section

	------------------------------------CONDITION NAMES

that can be filled by name,number pairs by the exodus preprocessor.
If it exists, the names can be used instead plain numbers at the beginning
of your condition lines.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4800 060e22d8-9d1c-0410-868a-e3a950d63d95