Add a very short documentation on the overall baci design.
This turns out to be just pointers to doxygen and a few hints on why things
are the way they are. It's a start, but I do not think we need much more.
Source code documentation comes from doxygen.

Please have a look and tell me what you think is missing.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4816 060e22d8-9d1c-0410-868a-e3a950d63d95