fix of nightly test cases for contact by including initialization of ConstraintManager and SurfStressManager into contactstrugenalpha constructor.
IMPORTANT NOTE:
at the moment the contactstrugenalpha class is a separate, derived class from strugenalpha (for development purposes), so if you make changes to the strugenalpha class (especially to the constructor) please copy/paste them at the same point into the contactstrugenalpha constructor (this would help me a lot)! You can also ask me, if you are not sure! Thanks.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4838 060e22d8-9d1c-0410-868a-e3a950d63d95