Add draft of system matrix wrapper with the ability to assemble block systems of any size.
This is a partial commit of unfinished work. The SystemMatrix wrapper is meant to be
used within baci algorithms instead of a plain CrsMatrix. But modifications to the
framework are not finished yet. Documentation is totally missing. However, the major
work of the column communication is there, so better commit this file right now.

To be continued.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4849 060e22d8-9d1c-0410-868a-e3a950d63d95