Introduce LINALG::SparseMatrix to all algorithms.
For the most part LINALG::SparseMatrix replaces Epetra_CrsMatrix. There is
not logical change involved right now. The only point of interesst is that
you will not reallocate a LINALG::SparseMatrix object but instead call its
Zero() method to clear it for reassembly.

The contact code needs to be updated. The solver bindings have not been
touched yet.

To be continued.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4941 060e22d8-9d1c-0410-868a-e3a950d63d95