STRESS OUTPUT FOR SOLIDS (PART I)
enabled output of nodal (!) xyz-stresses for SolidHex8. gauss point stresses are extrapolated to nodes, where an averaged value (in case of several adjacent elements) is written.

currently supported only for dynamics and GiD, but statics and ensight will follow.
stress output for other elements can easily be implemented accordingly, the only difference is in the matrix used for extrapolation of stresses from gauss points to nodes. 


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4962 060e22d8-9d1c-0410-868a-e3a950d63d95