Ported assembly of MFSI block matrix to BlockSparseMatrix class
There are internal matrix manipulations that belong neither to the SparseMatrix class itself
nor are part of the regular usage of the SparseMatrix interface. Right now these things are
done using an Epetra_CrsMatrix that is wrapped in a SparseMatrix afterwards. To accomplish
this s SparseMatrix object can be constructed via a shallow copy (view) of the original matrix.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5116 060e22d8-9d1c-0410-868a-e3a950d63d95