STRESS OUTPUT PART III
enabled GiD output of Gauss point stresses -> here no extrapolation to nodes or element center necessary
since GiD is capable of visualizing Gauss point quantities (in GiD 8 both discontinuous and smoothed).
Stress output depends (as everything in the drt GiD filter) on the assumption that only one element type 
is present, so no hybrid meshes supported.

I did it exemplarily for SolidHex8, but other elements can be implemented straightforwardly, too.  

NOTE: 
If the definition of Gauss points for an element in baci conflicts with that of GiD (see
http://gid.cimne.upc.es/support_team/gid_toc/gid_toc.html), a mapping is necessary (e.g. in case of wall
elements). For SolidHex8 this was not required since the convention is the same (thanks to Moritz). 


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5118 060e22d8-9d1c-0410-868a-e3a950d63d95