improvements for 2d contact:
(1) modified the way interfaces (slave and master sides) are built. Now an
interface is no longer restricted to a pair of contact conditions but can
be defined by arbitrarily large groups of conditions! This means that now the slave or master side of an interface may for example be defined in GID by more than one design line.

(2) included treatment of the special case of exact 1:1 contact node
projections (e.g. the projection of a slave node falls exactly onto a master node) into the overlap detection routine.

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5127 060e22d8-9d1c-0410-868a-e3a950d63d95