***!!!This new feature defaults to "ON" meaning if you don't have the appropriate flags in your input file you ARE using it and you ARE using the  default settings (which where chosen conservatively though)!!!***
Added functionality to adaptively control the linear solver tolerance when using an iterative linear solver. Currently hooked up to nonlinear fluid dynamics only but general feature lives in LINALG::Solver, so hookup to nonlinear structural dynamics will follow.

On how to control this feature I'll write a separate email to 'wir'


Note that depending on your choice of Aztec options it might yell at you with an dserror - ask me or wait for my email to 'wir' which will explain more.

This commit was presented to you by the 
"Fluid Efficiency in Baci Initiative (FEIBI)"   :o]

git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5132 060e22d8-9d1c-0410-868a-e3a950d63d95