Have all dofs in a problem be unique, even across multiple discretizations.
Unique dof numbers are needed if different discretizations contribute to a single block matrix.
(The block matrix has to have a map.) We keep a list of all DofSets here so that each DofSet
can ask its precessor for its largest global id. We rely on the fact that dof numbers are
assigned in the order DofSets are created. This is somewhat implicit and will be changed.

Amesos needs global matrices with continuous maps starting with 0. These can be created
easily using EpetraExt::LinearProblem_Reindex objects.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5146 060e22d8-9d1c-0410-868a-e3a950d63d95