Make sure the filter FillComplete()s its Discretizations in the same order as baci itself.
Warnung: We have very subtil dependencies here. In multi-Discretization problems (FSI)
all dof numbers are unique. This is achived implicitely in Discretization::FillComplete().
And we have to make sure the dof numbers used by the filters are the same as those used
by baci. So we have to make sure we FillComplete() our Discretizations in the right order.

The order the Discretizations are treated in the filters is determined by the control file.
We rely on our control files to be correct, that is we assume the Discretizations are
written in the same order as they have been filled.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5224 060e22d8-9d1c-0410-868a-e3a950d63d95