Combine AMD LAPACK (acml) with goto BLAS to obtain fastest version.
This uses libacml_goto, a patched up library that contains all acml
object files and all goto object files, where goto overwrites acml. This
might work well as blas (and goto and acml) put one function in one
file. The blas interfaces are of course clearly defined. The file names
are the same in both libraries... Still this is a test.

Unfortunatelly it seems that ATLAS cannot be combined with goto. And
ATLAS does not provide enough of LAPACK to satisfy Trilinos. (Even if we
do not use these parts.)


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5227 060e22d8-9d1c-0410-868a-e3a950d63d95