add experimental element group
The element group is used by our fluid algorithm and finds axis aligned hex8 elements
that could be evaluated much faster than ordinary (general) fluid elements. The element
evaluation is done by a special Evaluator, that represents exactly one element action
for the specified group of elements. The whole point of this design is to be able to
gain element evaluation speed.

Right now the basic framework is there. Real speed improvements are yet to come.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5253 060e22d8-9d1c-0410-868a-e3a950d63d95