OUTPUT OF STRUCTURAL STRAINS PART I
Enabled output of Green-Lagrange strains in GiD (Gauss point values) and 
ensight/paraview (extrapolated and averaged nodal or element center values).
You can switch on writing Green-Lagrange strains at the Gauss points
in your input file by choosing

STRUCT_STRAIN     Yes

Calculation of strains is coupled to calculation of stresses because it is
assumed that we do not want output of strains unless we also want the
corresponding stresses. Apart from that, it did not seem reasonable to
implement a new block calc_struct_strains doing almost the same as 
calc_struct_stress. When calculating the stresses, we can easily also
return the corresponding Green-Lagrange strains (probably enhanced due to
EAS). Thus in order to write structural strains, you must also choose

STRUCT_STRESS     Yes

If anyone does not like that coupling for a good reason, we can still 
change that easily.
If the calculated strains are then also written to BINIO depends on your
choice of STRUCT_STRAIN. 

Any further post-processing (like extrapolation to the nodes or calculation
of center strains) is done in the ensight-filter analogously to the
stresses. In order to reuse as most of the routines as possible, I sometimes
left the naming of the temporary variables ("stress" etc.) although these
routines can now also be used for strains.

I have implemented this feature for all elements currently supporting
the new stress output, but it was only tested for SOLID3 HEX8, so please
make extensive use of it and report any bugs to me.
GiD output only works for SOLID3 HEX8.

An update of lnm_general as well as the static control routine follows
immediately.



git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5300 060e22d8-9d1c-0410-868a-e3a950d63d95