OUTPUT OF STRUCTURAL STRAINS PART II
Enabled calculation of Green-Lagrange strains now also for static problems.
For that purpose got rid of the C structure ioflags and replaced it with
the corresponding Teuchos::ParameterList of the Problem. On this occasion
also removed the (useless) C structure from the structural dynamics part.

REMARK: The options for the ensight-filter for the strain output are the
following

     --straintype=cxyz          (for element center strains)
or 
     --straintype=ndxyz         (for nodal strains)
or  
     --straintype=cxyz_ndxyz    (for both)

just as in case of stresses. Of course you can choose stress _and_ strain 
output in any possible combination, so e.g.

--stresstype=cxyz_ndxyz --straintype=ndxyz

Currently only stresses and strains in the xyz-reference frame are supported.
Principal stresses/strains will follow.


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5302 060e22d8-9d1c-0410-868a-e3a950d63d95