Replaced Epetra_SerialDenseVector and Epetra_SerialDenseMatrix with Blitz Arrays for small vectors (length 3) or small matrices (3x3 or 3x2) in the intersection routines.
If sizes of blitz Arrays are fixed at compile time (which is the case, if these vectors hold e.g. coordinates), one sees nice speedups.

The open question would be, how pure C arrays compare speedwise - but then we loose the nice syntax of blitz, so it might be a tradeoff...


git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@5312 060e22d8-9d1c-0410-868a-e3a950d63d95